Config Key Refactor
This patch refactors the way configuration keys are handled such that
all keys are now constant, unsigned 32-bit integers typed as a
`ConfigKey`. This type can be used to obtain the key's path,
description, and other information. Additionally, there are methods in
`./api/types` named `ParentConfigKeys`, `ValueConfigKeys`, and
`GetConfigSectionInfo` that enable the retrieval of all parent
ConfigKeys, all value ConfigKeys, and the ability to query a key for its
value keys as well as any child, parent keys it may contain.