TOML Optional
This patch updates `glide.yaml` to override the transitive dependency
upon `github.com/spf13/viper` to point to a forked branch that removes
TOML support by default, and includes it only if the build tag `toml` is
supplied. This is to remove the dependency upon the library currently
used to parse TOML as it has a license that is incongruent with the K8
project's requirements.